feat: Complete major fixes and improvements
- Remove IncomingCallDialog and use only native CallKit/VoIP
- Fix CallKit notifications in foreground for Android
- Fix favorites loading error and improve UI design
- Fix FCM token cleanup and duplicate handling
- Fix VoIP token management and cleanup
- Fix notification payload inconsistencies
- Generate comprehensive CODING_RULES.md analysis (91+ violations)
- Create refactoring documentation (5 files, 69KB)

+++ lib/services/audio_processing_service.dart
@@ -27,13 +27,13 @@ class AudioProcessingService {
           return _generateDefaultWaveform();
         }
 
-        // Convertir a valores absolutos
-        final absValues = waveData.map((v) => v.abs()).toList();
+        // Convertir a valores absolutos y asegurar que sean double
+        final absValues = waveData.map((v) => v.abs().toDouble()).toList();
 
         // Encontrar min y max para normalización
         final maxVal = absValues.reduce((a, b) => a > b ? a : b);
         final minVal = absValues.reduce((a, b) => a < b ? a : b);
-        final range = maxVal - minVal;
+        final range = (maxVal - minVal).toDouble();
 
         if (range == 0) {
           print('⚠️ [AUDIO PROCESSING] Range es 0, usando datos por defecto');
